Fix title before name and suffix after name
English version is clear: Mr./Mrs. for prefix and Jr., Sr. for suffix.
So it's not that much related to university diploma.

This is important as when a html form require prefix, it must be clear,
that we mean by "title" Mr. and not university diploma (although it can
be).

Signed-off-by: Petr Vorel <petr.vorel@gmail.com>